Replace % formatting with format function.
The % operator for formatting was getting to confusing as the modulo operation
is a valid with strings too, and lists can be used for modulo as well. The new
format function is less confusing and has the advantage of better error
position reporting of parameter errors. The % operator only does modulo now.

Lists can still be used as parameters by splitting them manually to items as in
format("%d: %s", i[0], i[1]) or even better by using the splat operator as in
format("%d: %s", *i).

Old code like "%d: %s" % (1, "text") needs to be converted to format("%d: %s",
1, "text").

Sorry about this late change, the % operator was a bad choice. Some languages
use it for formatting, but those are not converting strings to integers.